Extract more CLI-arg-parsing code into new module
